transfer presenter refactoring
